feat: restructure Component order
