+++ frontend/components/YourWalletPage/styles.ts
